@@ -22,36 +22,36 @@
   },
   "private": true,
   "dependencies": {
-    "@angular-devkit/schematics": "^10.0.5",
-    "@angular/animations": "^10.0.8",
-    "@angular/cdk": "^9.2.0",
-    "@angular/common": "^10.0.8",
-    "@angular/compiler": "^10.0.8",
-    "@angular/core": "^10.0.8",
-    "@angular/flex-layout": "^9.0.0-beta.29",
-    "@angular/forms": "^10.0.8",
-    "@angular/http": "^7.2.16",
-    "@angular/material": "^9.2.0",
-    "@angular/platform-browser": "^10.0.8",
-    "@angular/platform-browser-dynamic": "^10.0.8",
-    "@angular/router": "^10.0.8",
-    "@microsoft/signalr": "^3.1.6",
-    "@ngrx/effects": "^9.0.0",
-    "@ngrx/store": "^9.0.0",
-    "caniuse-lite": "1.0.30001038",
+    "@angular-devkit/schematics": "^10.2.0",
+    "@angular/animations": "^10.2.3",
+    "@angular/cdk": "^10.2.7",
+    "@angular/common": "^10.2.3",
+    "@angular/compiler": "^10.2.3",
+    "@angular/core": "^10.2.3",
+    "@angular/flex-layout": "^10.0.0-beta.32",
+    "@angular/forms": "^10.2.3",
+    "@angular/material": "^10.2.7",
+    "@angular/platform-browser": "^10.2.3",
+    "@angular/platform-browser-dynamic": "^10.2.3",
+    "@angular/router": "^10.2.3",
+    "@caiu/library": "^10.3.6",
+    "@microsoft/signalr": "^3.1.10",
+    "@ngrx/effects": "^9.2.0",
+    "@ngrx/store": "^9.2.0",
+    "ngx-color-picker": "^10.1.0",
     "rxjs": "~6.5.4",
-    "tinymce": "^5.2.1",
-    "tslib": "^2.0.0",
+    "tinymce": "^5.6.2",
+    "tslib": "^2.0.1",
     "zone.js": "~0.10.3"
   },
   "devDependencies": {
     "@angular-devkit/build-ng-packagr": "~0.1000.5",
     "ng-packagr": "^10.0.0",
     "tsickle": ">=0.38.1",
-    "@angular-devkit/build-angular": "^0.1000.5",
-    "@angular/cli": "^10.0.5",
-    "@angular/compiler-cli": "^10.0.8",
-    "@angular/language-service": "^10.0.8",
+    "@angular-devkit/build-angular": "^0.1002.0",
+    "@angular/cli": "^10.2.0",
+    "@angular/compiler-cli": "^10.2.3",
+    "@angular/language-service": "^10.2.3",
     "@types/jasmine": "~3.5.10",
     "@types/jasminewd2": "~2.0.8",
     "@types/node": "~13.9.5",
