@@ -3,7 +3,7 @@ import { Observable } from 'rxjs';
 
 import { HttpService } from './http.service';
 import { HttpGetPayload, HttpPostPayload, HttpPutPayload, HttpDeletePayload } from './http.models';
-import { Command } from '../shared/decorators';
+// import { Command } from '../shared/decorators';
 
 @Injectable({
   providedIn: 'root'
@@ -13,27 +13,27 @@ export class HttpCommands<T> {
   constructor(private http: HttpService) {
   }
 
-  @Command()
+  // @Command()
   delete$(payload: HttpDeletePayload<T>): Observable<any> {
     return this.http.delete(payload.path, payload.headers, payload.options);
   }
 
-  @Command()
+  // @Command()
   get$(payload: HttpGetPayload<T>): Observable<any> {
     return this.http.get(payload.path, payload.headers, payload.options);
   }
 
-  @Command()
+  // @Command()
   post$(payload: HttpPostPayload<T>): Observable<any> {
     return this.http.post(payload.path, payload.model, payload.headers, payload.options);
   }
 
-  @Command()
+  // @Command()
   postFormUrlEncoded$(payload: HttpPostPayload<T>): Observable<any> {
     return this.http.postFormUrlEncoded(payload.path, payload.model, payload.headers, payload.options);
   }
 
-  @Command()
+  // @Command()
   put$(payload: HttpPutPayload<T>): Observable<any> {
     return this.http.put(payload.path, payload.model, payload.headers, payload.options);
   }
